# core/API-swagger.yaml
openapi: 3.1.0
info:
  title: SupremeAI 2.0
  version: 0.1.0
paths:
...
  /marketplace/search:
    post:
      tags:
      - marketplace
      summary: Search Marketplaces
      operationId: search_marketplaces_marketplace_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /marketplace/install:
    post:
      tags:
      - marketplace
      summary: Install Tool
      operationId: install_tool_marketplace_install_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api__routes__marketplace_endpoints__InstallRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SearchRequest:
      properties:
        query:
          type: string
          title: Query
        categories:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Categories
        filters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Filters
        installed_only:
          type: boolean
          title: Installed Only
          default: false
      type: object
      required:
      - query
      title: SearchRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    api__routes__marketplace_endpoints__InstallRequest:
      properties:
        tool_id:
          type: string
          title: Tool Id
        target_environment:
          type: string
          title: Target Environment
          default: local
        sandbox:
          type: boolean
          title: Sandbox
          default: true
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
      type: object
      required:
      - tool_id
      title: InstallRequest
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login